requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==1.0.4
pydantic==2.5.3
python-dotenv==1.0.0
armoriq-sdk==0.2.6
google-generativeai==0.3.2
